requirements.txt
absl-py==0.12.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
antlr4-python3-runtime==4.8
anykeystore==0.2
apex==0.1
async-timeout==4.0.1
attrs==21.2.0
axial-positional-embedding==0.2.1
blobfile==1.2.7
braceexpand==0.1.7
cachetools==4.2.2
certifi==2020.12.5
chainer==7.8.1
chainercv==0.13.1
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.1
clip @ git+https://github.com/openai/CLIP.git@cfcffb90e69f37bf2ff1e988237a0fbe41f33c04
cmapy==0.6.6
cryptacular==1.6.2
cycler==0.10.0
Cython==0.29.23
DALL-E==0.1
dalle-pytorch==1.1.4
dataclasses==0.6
decorator==4.4.2
deepspeed==0.5.7
defusedxml==0.7.1
dominate==2.6.0
easydict==1.9
efficientnet-pytorch==0.7.1
einops==0.3.2
filelock==3.4.0
frozenlist==1.2.0
fsspec==2021.11.1
ftfy==6.0.3
future==0.18.2
g-mlp-pytorch==0.1.5
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grad-cam==1.2.3
graphviz==0.8.4
greenlet==1.1.2
grpcio==1.37.1
hjson==3.0.2
huggingface-hub==0.1.2
hupper==1.10.3
idna==2.10
imageio==2.9.0
imgaug==0.4.0
iniconfig==1.1.1
inplace-abn==1.1.0
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
mmcv-full==1.4.3
# Editable install with no version control (mmsegmentation==0.20.2)
-e /data1/xjheng/workspace/programs/ECCV2022/mmsegmentation
multidict==5.2.0
mxnet @ file:///data1/xjheng/packages/mxnet-1.7.0.post2-py2.py3-none-manylinux2014_x86_64.whl
mypy==0.910
mypy-extensions==0.4.3
networkx==2.5.1
ninja==1.10.2.3
numpy==1.20.2
oauthlib==3.1.0
omegaconf==2.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
packaging==21.3
pandas==1.2.4
PasteDeploy==2.1.1
pbkdf2==1.3
Pillow==8.2.0
plaster==1.0
plaster-pastedeploy==0.7
pluggy==1.0.0
prettytable==3.0.0
protobuf==3.15.8
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycryptodomex==3.11.0
pydensecrf==1.0rc2
pyDeprecate==0.3.1
pyparsing==2.4.7
pyramid==2.0
pyramid-mailer==0.15.1
pytest==6.2.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytorch-lightning==1.5.3
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
randaugment==1.0.2
regex==2021.4.4
repoze.sendmail==4.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resnest==0.0.6b20210715
rotary-embedding-torch==0.1.2
rsa==4.7.2
sacremoses==0.0.46
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.2
Shapely==1.7.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.25
taming-transformers-rom1504==0.0.6
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
threadpoolctl==2.1.0
tifffile==2021.4.8
timm==0.4.12
tokenizers==0.10.3
toml==0.10.2
torch==1.7.1+cu110
torchfile==0.1.0
torchmetrics==0.6.0
torchnet==0.0.4
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.60.0
transaction==3.0.1
transformers==4.12.5
translationstring==1.4
triton==1.1.1
tsnecuda==0.1.1
typing-extensions==3.7.4.3
urllib3==1.26.4
velruse==1.1.1
venusian==3.0.0
visdom==0.1.8.9
wcwidth==0.2.5
webdataset==0.1.103
WebOb==1.8.7
websocket-client==1.1.0
Werkzeug==1.0.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
xmltodict==0.12.0
yapf==0.32.0
yarl==1.7.2
youtokentome==1.0.6
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
